Мерж из 0.2.1
git-svn-id: https://horse.aplana.com/svn/SBRFACCTAX/trunk/taxaccounting@2057 34aa7802-8fb0-764d-beb8-9b5b9d595954

+++ dao/src/main/java/com/aplana/sbrf/taxaccounting/dao/impl/DepartmentParamDaoImpl.java
@@ -80,7 +80,7 @@ public class DepartmentParamDaoImpl extends AbstractDao implements DepartmentPar
 							departmentParamIncome.setSignatoryLastName(rs.getString("SIGNATORY_LASTNAME"));
 							departmentParamIncome.setSignatorySurname(rs.getString("SIGNATORY_SURNAME"));
 							departmentParamIncome.setTaxPlaceTypeCode(rs.getString("TAX_PLACE_TYPE_CODE"));
-							departmentParamIncome.setTaxRate(Long.valueOf(rs.getLong("TAX_RATE")));
+							departmentParamIncome.setTaxRate(rs.getDouble("TAX_RATE"));
 							departmentParamIncome.setSumDifference(Long.valueOf(rs.getLong("SUM_DIFFERENCE")));
 							departmentParamIncome.setAppVersion(rs.getString("APP_VERSION"));
 							departmentParamIncome.setFormatVersion(rs.getString("FORMAT_VERSION"));
